# swagger/swagger.yaml
basePath: /api/v0
definitions:
...
  consoleapi.AddProjectMemberACLBucketSwaggerRequest:
    properties:
      bucketName:
        example: gmail
        type: string
    type: object
...
  consoleapi.DeleteMembersAndInvitationsSwaggerRequest:
    properties:
      emails:
        example:
        - member@example.com
        - pending@example.com
        items:
          type: string
        type: array
      removeAccesses:
        example: true
        type: boolean
    type: object
...
  consoleapi.MemberBucketGrantInputSwagger:
    properties:
      allowDelete:
        example: false
        type: boolean
      allowDownload:
        example: true
        type: boolean
      allowList:
        example: true
        type: boolean
      allowUpload:
        example: false
        type: boolean
      bucket:
        example: gmail
        type: string
      prefix:
        example: member@example.com/
        type: string
    type: object
  consoleapi.MemberBucketGrantSwaggerItem:
    properties:
      allowDelete:
        example: false
        type: boolean
      allowDownload:
        example: true
        type: boolean
      allowList:
        example: true
        type: boolean
      allowUpload:
        example: false
        type: boolean
      bucket:
        example: gmail
        type: string
      createdAt:
        type: string
      id:
        example: 00000000-0000-0000-0000-000000000000
        type: string
      inviteEmail:
        example: member@example.com
        type: string
      memberId:
        example: 00000000-0000-0000-0000-000000000001
        type: string
      prefix:
        example: member@example.com/
        type: string
      projectId:
        example: 00000000-0000-0000-0000-000000000000
        type: string
      updatedAt:
        type: string
    type: object
...
  consoleapi.ProjectInvitationSwaggerItem:
    properties:
      createdAt:
        type: string
      email:
        example: invitee@example.com
        type: string
      expired:
        example: false
        type: boolean
    type: object
  consoleapi.ProjectMemberACLBucketSwaggerItem:
    properties:
      bucketName:
        example: gmail
        type: string
      createdAt:
        type: string
      projectId:
        example: 00000000-0000-0000-0000-000000000000
        type: string
    type: object
  consoleapi.ProjectMemberDetailSwaggerResponse:
    properties:
      id:
        example: 00000000-0000-0000-0000-000000000001
        type: string
      joinedAt:
        type: string
      projectID:
        example: 00000000-0000-0000-0000-000000000000
        type: string
      role:
        example: 1
        type: integer
    type: object
  consoleapi.ProjectMemberSwaggerItem:
    properties:
      email:
        example: jane@example.com
        type: string
      fullName:
        example: Jane Doe
        type: string
      id:
        example: 00000000-0000-0000-0000-000000000001
        type: string
      joinedAt:
        type: string
      role:
        example: 1
        type: integer
      shortName:
        example: Jane
        type: string
    type: object
  consoleapi.ProjectMembersPageSwaggerResponse:
    properties:
      currentPage:
        example: 1
        type: integer
      limit:
        example: 100
        type: integer
      offset:
        example: 0
        type: integer
      order:
        example: 1
        type: integer
      orderDirection:
        example: 1
        type: integer
      pageCount:
        example: 1
        type: integer
      projectInvitations:
        items:
          $ref: '#/definitions/consoleapi.ProjectInvitationSwaggerItem'
        type: array
      projectMembers:
        items:
          $ref: '#/definitions/consoleapi.ProjectMemberSwaggerItem'
        type: array
      search:
        example: ""
        type: string
      totalCount:
        example: 3
        type: integer
    type: object
...
  consoleapi.PutMemberBucketGrantsSwaggerRequest:
    properties:
      grants:
        items:
          $ref: '#/definitions/consoleapi.MemberBucketGrantInputSwagger'
        type: array
    type: object
...
  consoleapi.SwaggerErrorResponse:
    properties:
      error:
        example: request validation failed
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API documentation for StorX Monitor server
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: StorX Monitor API
  version: "1.0"
paths:
...
  /projects/{id}/member-acl-buckets:
    get:
      description: '**Full route:** `GET /api/v0/projects/{id}/member-acl-buckets`'
      parameters:
      - description: Project UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/consoleapi.ProjectMemberACLBucketSwaggerItem'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
      security:
      - CookieAuth: []
      summary: '[3] List ACL bucket registry'
      tags:
      - member-bucket-restriction
    post:
      consumes:
      - application/json
      description: '**Full route:** `POST /api/v0/projects/{id}/member-acl-buckets`'
      parameters:
      - description: Project UUID
        in: path
        name: id
        required: true
        type: string
      - description: Bucket to register
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/consoleapi.AddProjectMemberACLBucketSwaggerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/consoleapi.ProjectMemberACLBucketSwaggerItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
      security:
      - CookieAuth: []
      - CSRFAuth: []
      summary: '[3] Register bucket for Member ACL'
      tags:
      - member-bucket-restriction
  /projects/{id}/member-acl-buckets/{bucket}:
    delete:
      description: '**Full route:** `DELETE /api/v0/projects/{id}/member-acl-buckets/{bucket}`'
      parameters:
      - description: Project UUID
        in: path
        name: id
        required: true
        type: string
      - description: Bucket name
        in: path
        name: bucket
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
      security:
      - CookieAuth: []
      - CSRFAuth: []
      summary: '[3] Unregister ACL bucket'
      tags:
      - member-bucket-restriction
  /projects/{id}/members:
    delete:
      consumes:
      - application/json
      description: '**Full route:** `DELETE /api/v0/projects/{id}/members`'
      parameters:
      - description: Project UUID
        in: path
        name: id
        required: true
        type: string
      - description: Emails to remove
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/consoleapi.DeleteMembersAndInvitationsSwaggerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
      security:
      - CookieAuth: []
      - CSRFAuth: []
      summary: '[8] Remove members / cancel invites'
      tags:
      - member-bucket-restriction
    get:
      description: '**Full route:** `GET /api/v0/projects/{id}/members`'
      parameters:
      - description: Project UUID
        in: path
        name: id
        required: true
        type: string
      - default: 100
        description: Page size
        in: query
        name: limit
        required: true
        type: integer
      - default: 1
        description: Page number (1-based)
        in: query
        name: page
        type: integer
      - description: Search by name/email
        in: query
        name: search
        type: string
      - default: 1
        description: 'Sort field: 1=name, 2=email, 3=created'
        in: query
        name: order
        type: integer
      - default: 1
        description: 1=asc, 2=desc
        in: query
        name: order-direction
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/consoleapi.ProjectMembersPageSwaggerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
      security:
      - CookieAuth: []
      summary: '[5] List project members and pending invites'
      tags:
      - member-bucket-restriction
  /projects/{id}/members/{memberID}:
    get:
      description: '**Full route:** `GET /api/v0/projects/{id}/members/{memberID}`'
      parameters:
      - description: Project UUID
        in: path
        name: id
        required: true
        type: string
      - description: Member user UUID
        in: path
        name: memberID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/consoleapi.ProjectMemberDetailSwaggerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
      security:
      - CookieAuth: []
      summary: '[8] Get one project member'
      tags:
      - member-bucket-restriction
    patch:
      consumes:
      - text/plain
      description: '**Full route:** `PATCH /api/v0/projects/{id}/members/{memberID}`'
      parameters:
      - description: Project UUID
        in: path
        name: id
        required: true
        type: string
      - description: Member user UUID
        in: path
        name: memberID
        required: true
        type: string
      - description: 'Role: 0=Admin, 1=Member'
        example: 1
        in: body
        name: body
        required: true
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/consoleapi.ProjectMemberSwaggerItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
      security:
      - CookieAuth: []
      - CSRFAuth: []
      summary: '[8] Update member role (Admin/Member)'
      tags:
      - member-bucket-restriction
  /projects/{id}/members/{memberID}/bucket-grants:
    get:
      description: '**Full route:** `GET /api/v0/projects/{id}/members/{memberID}/bucket-grants`'
      parameters:
      - description: Project UUID
        in: path
        name: id
        required: true
        type: string
      - description: Member user UUID
        in: path
        name: memberID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/consoleapi.MemberBucketGrantSwaggerItem'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
      security:
      - CookieAuth: []
      summary: '[7] Get Member folder grants'
      tags:
      - member-bucket-restriction
    put:
      consumes:
      - application/json
      description: '**Full route:** `PUT /api/v0/projects/{id}/members/{memberID}/bucket-grants`'
      parameters:
      - description: Project UUID
        in: path
        name: id
        required: true
        type: string
      - description: Member user UUID
        in: path
        name: memberID
        required: true
        type: string
      - description: Full grant set (empty clears access)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/consoleapi.PutMemberBucketGrantsSwaggerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/consoleapi.MemberBucketGrantSwaggerItem'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
      security:
      - CookieAuth: []
      - CSRFAuth: []
      summary: '[7] Replace Member folder grants'
      tags:
      - member-bucket-restriction
...
securityDefinitions:
  ApiKeyAuth:
    description: Bearer / API key (rare for console routes).
    in: header
    name: Authorization
    type: apiKey
  CSRFAuth:
    description: 'When `console.csrf-protection-enabled` is true: copy `csrfToken`
      from `GET /config` and Authorize here. Must match `csrf_token` cookie (set by
      GET /config or synced by Swagger UI). Required for `POST /auth/token`, `POST
      /auth/account/set-password`, and other CSRF-protected routes.'
    in: header
    name: X-CSRF-Token
    type: apiKey
  CookieAuth:
    description: Session cookie after login. **Swagger Authorize:** paste the token
      value only (from `POST /auth/token` response `token` field). Do not prefix with
      `Cookie:` or `_tokenKey=`. Copied curl uses `-b '_tokenKey=...'` (not a header).
      Login via Swagger auto-fills this.
    in: cookie
    name: _tokenKey
    type: apiKey
  SellerCSRFAuth:
    description: 'When `seller.csrf-protection-enabled` is true: copy `csrfToken`
      from `GET /seller/config` and Authorize here. Must match `csrf_token` cookie
      (set by GET /seller/config). Required for `POST /seller/auth/token`, MFA routes,
      password change, and other CSRF-protected seller routes.'
    in: header
    name: X-CSRF-Token
    type: apiKey
  SellerCookieAuth:
    description: Session cookie after reseller Google login. **Swagger Authorize:**
      paste the token value only (from `GET /seller/auth/google` response `token`
      field). Seller peer cookie name is `_seller_tokenKey` (not console `_tokenKey`).
    in: cookie
    name: _seller_tokenKey
    type: apiKey
swagger: "2.0"
...